Workaround for CMake
In CMake, subdirectories inherit the variables of the parent directory.
This is a problem for us because it means that if we give -DBUILD_TESTS=1,
then this will be inherited to the SBWT subdirectory, which will also then
enable tests, but then googletest is included twice and CMake crashes.
So either we need to find a way to prevent the child from inheriting
the variable, or we need to rename the variable in the parent. Googling
around for the first option for 15 minutes did not yield a way to do it
(seriously, fuck CMake), so we just rename the variable.