Can't build tests and main executable at the same time anymore
Some code that interfaces with KMC in SBWT was moved out from headers into
source files. The implication of this is that the max k is now fixed at
SBWT. This means that we can't build themisto tests and the themisto main
executable in the same build anymore because they require a different
maximum value of k (255 at tests). This is not an impossible obstacle to
overcome, but it requires some work, so for now we just disable the build
target of the main exucutable if building tests.