Fix performance issue.
The assignment operator of color set from another color set copied
the whole data pointed by the other set, not just the relevant slice.
This meant that if the data pointer was to the concatenation of all
color sets, the whole concatenation was copied over. This was not an issue
until two commits ago, because that scenario never happened.