Fix space bug if there is only 1 color in the whole data.
The bit width of color ids was set to ceil(log2(number_of_colors)).
But if the number of colors is 1, then that expression is equal to 0.
I think the effect of this is that the sdsl integer vector uses full 64
bits for each integer, which makes the space 64 times larger than necessary.

Fixed by setting the bit width to 1 if there is just 1 color.